% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{ial_district}
\alias{ial_district}
\title{Time series for a single district}
\usage{
ial_district(district_id, interval_start, interval_stop)
}
\arguments{
\item{district_id}{a length 1 character vector with a single district ID}

\item{interval_start, interval_end}{start/end dates for itervals. They can
be \code{Date} objects (only the year and month will be used in the API
call) or strings. If strings, they can be full ISO date strings
(i.e. "2018-09-29") or strings in the \code{yy.mm} format the API
described. They will be handled automagically for you but the
function assumes that \code{interval_start} and \code{interval_end} both
have the same format string or are both \code{Date} objects.}
}
\value{
data frame
}
\description{
Time series for a single district
}
