% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-regions.R
\name{ial_regions}
\alias{ial_regions}
\title{List of all the regions (districts) with display name and id}
\usage{
ial_regions()
}
\value{
data frame
}
\description{
List of all the regions (districts) with display name and id
}
