% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indialights-package.R
\docType{package}
\name{indialights}
\alias{indialights}
\alias{indialights-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/indialights}
\item BugReports: \url{https://gitlab.com/hrbrmstr/indialights/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
