% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indialights-package.R
\docType{package}
\name{indialights}
\alias{indialights}
\alias{indialights-package}
\title{Tools to Work with the 'India Lights' 'API'}
\description{
The 'India Lights' 'API' shows light output at night for 20 years, from 1993
to 2013, for 600,000 villages across India. The 'Defense Meteorological
Satellite Program' ('DMSP'), run by the 'U.S.' 'Department of Defense', has
taken pictures of the Earth every night for 20 years. Researchers at the
'University of Michigan' used the 'DMSP' images to extract the data and
provide it in tabular form. The 'India Lights' 'API' provides the data at
convenient endpoints that allows you to look at specific time intervals and
administration levels. Tools are provided to query these 'API' endpoints.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/indialights}
\item BugReports: \url{https://gitlab.com/hrbrmstr/indialights/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
